THE BIG REFACTORING - Episode I
Ok, some of the assumptions of my initial code were ~wrong~ too closely tight to the fact that a new game starts with the classic chess starting position.
We can now start games with any FEN, without having too many hard-coded bits related to that classic chess starting position.

Also, promotions were not managed. We now add a 3rd character to the piece's PieceRole to encode the promotion (hard-coded to be a Queen for the moment).