[gameplay] The depth the bot plays with can now be customised par challenge.
The depth we simulate the player's moves with can also be customised per challenge.

Future challenges will have a lower value for both values, as I realised that most of the challenges are too hard for me at the moment.
I'm a _very_ casual chess player, and I'm making this web game for people like me in that respect, so let's play it "easy mode" -  nothing to be ashamed of! ^_^