[lichess] Use "game stream" endpoint instead of the "game export" one
For some reason, it seems that the "game export" ednpoint returns outdated information about the ongoing games.

It was returning up-to-date data until yesterday, but for a mysterious reason it stopped doing so 🤷
Maybe it returns only the few first moves, when the game is still ongoing?

Anyway... Thankfully, it seems that using the "Stream Board game state" endpoint works correctly. 🤞🤞🤞

Also improve code there and there while I was here :-)