Switch the backend to a JavaScript one (Hono)
As I was about to start implementing the communication between that frontend and the Python backend, I realised that I was about to face the same issue I had in my very first version on Zakuchess, based on Django and Stimulus: a lot of duplicated logic between the frontend and the backend!

The only way to avoid that is to have a backend that is also written in TypeScript, so that both frontend and backend can share some chess-related and Zakuchess-related code.